clear filters before searching by tags or sources
